adjusted some types within datastructure
atomic mass float32 -> int
- a given atom has an integer mass. Various isotopes will be represented
through multiple different atom instances so mass will only ever be
of integer type

Vec2.X,Vec2.Y float32 -> float64
- better precision is always good with this kinda thing
- the relevant parts of standard library and ebiten mostly use 64 bit
rather than 32 bit so less conversions will be necessary

+++ main.go
@@ -50,7 +50,7 @@ func (g Game) Draw(screen *ebiten.Image) {
 	accelRegion := ebiten.NewImage(int(g.accelerationRegion.Width()), int(g.accelerationRegion.Height()))
 	accelRegion.Fill(color.RGBA{250, 50, 50, 100})
 	drawOps := ebiten.DrawImageOptions{}
-	drawOps.GeoM.Translate(float64(g.accelerationRegion.Min.X), float64(g.accelerationRegion.Min.Y))
+	drawOps.GeoM.Translate(g.accelerationRegion.Min.X, g.accelerationRegion.Min.Y)
 	screen.DrawImage(accelRegion, &drawOps)
 
 	g.drawableMethane.Draw(screen)
